* Fixed various GCC 4.0 and MacOS (tiger) issues

+++ packages/xpce/src/unx/file.c
@@ -1082,7 +1082,7 @@ an ISO Latin-1 string.
 status
 storeStringFile(FileObj f, String s)
 { if ( isstrA(s) )
-  { TRY(storeWordFile(f, (Any) s->size));
+  { TRY(storeWordFile(f, (Any) (long)s->size));
     Sfwrite(s->s_textA, sizeof(char), s->size, f->fd);
 
     DEBUG(NAME_save, Cprintf("Saved ISO string, %ld chars\n", s->size));
@@ -1090,7 +1090,7 @@ storeStringFile(FileObj f, String s)
   { const charW *w = s->s_textW;
     const charW *e = &w[s->size];
 
-    TRY(storeWordFile(f, (Any) s->size));
+    TRY(storeWordFile(f, (Any) (long)s->size));
     for( ; w<e; w++)
     { if ( Sputc(*w, f->fd) < 0 )
 	return checkErrorFile(f);
